add gps

+++ restaurants/DiscoverRestaurants/app/src/main/java/com/jtigernova/discoverrestaurants/repository/IRestaurantRepository.kt
@@ -7,5 +7,5 @@ import kotlinx.coroutines.CoroutineScope
 interface IRestaurantRepository {
     var scope: CoroutineScope?
 
-    fun getRestaurants(lat: Float, lng: Float): LiveData<List<Restaurant>>
+    fun getRestaurants(lat: Double, lng: Double): LiveData<List<Restaurant>>
 }
