+++ restaurants/DiscoverRestaurants/app/src/main/java/com/jtigernova/discoverrestaurants/repository/api/IRestaurantApi.kt
@@ -6,5 +6,5 @@ import kotlinx.coroutines.CoroutineScope
 interface IRestaurantApi {
     var scope: CoroutineScope?
 
-    suspend fun getRestaurants(lat: Float, lng: Float): List<Restaurant>
+    suspend fun getRestaurants(lat: Double, lng: Double): List<Restaurant>
 }
